global: choose dict approach
