+++ jsonalchemy/version.py
@@ -30,4 +30,4 @@ This file is imported by ``jsonalchemy.__init__``, and parsed by
 # Do not change the format of this next line. Doing so risks breaking
 # setup.py and docs/conf.py
 
-__version__ = "0.0.1.dev20150413"
+__version__ = "0.0.1.dev20150504"
