--- tests/schemas/compose/author.json
@@ -1,4 +1,0 @@
-{
-    "title": "Author compose fixture",
-    "type": "string"
-}
