--- tests/schemas/compose/title.json
@@ -1,4 +1,0 @@
-{
-    "title": "Title compose fixture",
-    "type": "string"
-}
