--- tests/schemas/creation_date.json
@@ -1,10 +1,0 @@
-{
-    "title": "Test creation date schema",
-    "type": "object",
-    "properties": {
-        "creation_date": {
-            "type": "string",
-            "format": "date"
-        }
-    }
-}
