--- tests/schemas/inherit/base.json
@@ -1,12 +1,0 @@
-{
-    "title": "Base inheritance schema",
-    "type": "object",
-    "properties": {
-        "title": {
-            "type": "string"
-        },
-        "author": {
-            "type": "string"
-        }
-    }
-}
