--- tests/schemas/inherit/title.json
@@ -1,17 +1,0 @@
-{
-    "title": "Title schema",
-    "type": "object",
-    "properties": {
-        "title": {
-            "type": "object",
-            "properties": {
-                "name": {
-                    "type": "string"
-                },
-                "subtitle": {
-                    "type": "string"
-                }
-            }
-        }
-    }
-}
