+++ tests/schemas/invalid_type.json
@@ -1,5 +1,5 @@
 {
-    "title": "Test simple schema",
+    "title": "Schema with invalid type",
     "type": "object",
     "properties": {
         "my_field": {
