--- tests/schemas/mixin/author.json
@@ -1,9 +1,0 @@
-{
-    "title": "Test author schema",
-    "type": "object",
-    "properties": {
-        "author": {
-            "type": "string"
-        }
-    }
-}
