--- tests/schemas/mixin/authors.json
@@ -1,21 +1,0 @@
-{
-    "title": "Test mixin authors schema",
-    "type": "object",
-    "properties": {
-        "authors": {
-            "type": "array",
-            "items": {
-                "type": "string"
-            }
-        },
-        "author": {
-            "type": "string"
-        },
-        "other_authors": {
-            "type": "array",
-            "items": {
-                "type": "string"
-            }
-        }
-    }
-}
