--- tests/schemas/mixin/keywords.json
@@ -1,21 +1,0 @@
-{
-    "title": "Test mixin keywords schema",
-    "type": "object",
-    "properties": {
-        "keywords": {
-            "type": "array",
-            "items": {
-                "type": "string"
-            }
-        },
-        "first_keyword": {
-            "type": "string"
-        },
-        "other_keywords": {
-            "type": "array",
-            "items": {
-                "type": "string"
-            }
-        }
-    }
-}
