--- tests/schemas/mixin/title.json
@@ -1,8 +1,0 @@
-{
-    "type": "object",
-    "properties": {
-        "title": {
-            "type": "string"
-        }
-    }
-}
