--- tests/schemas/mixin/titles.json
@@ -1,14 +1,0 @@
-{
-    "title": "Various kinds of titles.",
-    "type": "array",
-    "items": [
-        { "type": "string" },
-        {
-            "type": "object",
-            "properties": {
-                "title": { "type": "string" },
-                "subtitle": { "type": "string" }
-            }
-        }
-    ]
-}
